[#333] Enhance granularity of time structures throughout YottaDB
ABS_TIME, a structure used through the codebase, only allowed
granularity to within microseconds. This presented problems when
reporting durations that were much shorter than that, as reported by
customers in issue #333. This commit removes the ABS_TIME structure, and
instead typedef's it to the structures used by the C library.